// src/generated/types/DoctorDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type DoctorDto = {
  /**
   * @type string
   */
  id: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  username: string
  /**
   * @type string
   */
  email: string
  /**
   * @type string
   */
  title: string
  /**
   * @type string
   */
  department: string
  /**
   * @type boolean
   */
  status: boolean
  /**
   * @type string, date-time
   */
  lastLogin: Date
  /**
   * @description List of course codes the doctor teaches
   * @type array
   */
  assignedCourses: string[]
}