// src/generated/types/SoftwareListDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type SoftwareListDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  requiredMemory: string
  /**
   * @type string
   */
  requiredStorage: string
  /**
   * @type string
   */
  id: string
}